move type to src/constants/path.ts
