+++ src/components/svgs/lines/styles/jr-east-single-color.tsx
@@ -9,7 +9,7 @@ import {
     LineStyleComponentProps,
     LineStyleType,
 } from '../../../../constants/lines';
-import { OpenPath } from '../../../../util/path';
+import { OpenPath } from '../../../../constants/path';
 import {
     defaultJREastSingleColorDecorationAttributes,
     getJREastDecorationMarkerProps,
