+++ src/util/path.ts
@@ -1,147 +1,30 @@
-export interface PathPoint {
-    x: number;
-    y: number;
-}
+import {
+    ClosePath,
+    CubicTo,
+    LineTo,
+    LinearPath,
+    MoveTo,
+    OpenPath,
+    OpenPathDrawCommand,
+    Path,
+    PathCommand,
+    PathPoint,
+    RoundedTurnPath,
+    ShortOpenPath,
+    makeComplexOpenPath,
+    makeLinearPath,
+    makePoint,
+    makeRoundedTurnPath,
+    makeSharpTurnPath,
+} from '../constants/path';
 
-interface MoveTo {
-    cmd: 'M';
-    to: PathPoint;
-}
-
-interface LineTo {
-    cmd: 'L';
-    to: PathPoint;
-}
-
-interface CubicTo {
-    cmd: 'C';
-    c1: PathPoint;
-    c2: PathPoint;
-    to: PathPoint;
-}
-
-interface ClosePath {
-    cmd: 'Z';
-}
-
-type OpenPathDrawCommand = LineTo | CubicTo;
-type PathCommand = MoveTo | OpenPathDrawCommand | ClosePath;
-
-/**
- * Structured path model used across the renderer.
- *
- * `kind` is the coarse geometry classification consumed by style/path helpers,
- * while `commands` keeps the precise command sequence and `d` is the serialized SVG path data.
- */
-interface BasePath<C extends readonly PathCommand[]> {
-    readonly kind: string;
-    readonly commands: C;
-    readonly d: string;
-}
-
-/** Straight open path: `M L`. */
-export interface LinearPath extends BasePath<readonly [MoveTo, LineTo]> {
-    readonly kind: 'ml';
-}
-
-/** Sharp corner open path: `M L L`. */
-export interface SharpTurnPath extends BasePath<readonly [MoveTo, LineTo, LineTo]> {
-    readonly kind: 'mll';
-}
-
-/** Rounded corner open path: `M L C L`. */
-export interface RoundedTurnPath extends BasePath<readonly [MoveTo, LineTo, CubicTo, LineTo]> {
-    readonly kind: 'mlcl';
-}
-
-/** Catch-all for longer open paths that do not match the short-path special cases above. */
-export interface ComplexOpenPath
-    extends BasePath<readonly [MoveTo, OpenPathDrawCommand, OpenPathDrawCommand, ...OpenPathDrawCommand[]]> {
-    readonly kind: 'complex-open';
-}
-
-/** Closed filled geometry, typically derived from offset/outline helpers. */
-export interface ClosedAreaPath
-    extends BasePath<readonly [MoveTo, OpenPathDrawCommand, OpenPathDrawCommand, ...OpenPathDrawCommand[], ClosePath]> {
-    readonly kind: 'closed-area';
-}
-
-export type OpenPath = LinearPath | SharpTurnPath | RoundedTurnPath | ComplexOpenPath;
-
-export type Path = OpenPath | ClosedAreaPath;
-
-export type ShortOpenPath = LinearPath | SharpTurnPath | RoundedTurnPath;
-
-export const makePoint = (x: number, y: number): PathPoint => ({ x, y });
-
-export const moveTo = (to: PathPoint): MoveTo => ({ cmd: 'M', to });
-
-export const lineTo = (to: PathPoint): LineTo => ({ cmd: 'L', to });
-
-export const cubicTo = (c1: PathPoint, c2: PathPoint, to: PathPoint): CubicTo => ({ cmd: 'C', c1, c2, to });
-
-export const closePath = (): ClosePath => ({ cmd: 'Z' });
-
-const formatNumber = (value: number) => `${value}`;
-
-const serializeCommand = (command: PathCommand): string => {
-    switch (command.cmd) {
-        case 'M':
-        case 'L':
-            return `${command.cmd} ${formatNumber(command.to.x)} ${formatNumber(command.to.y)}`;
-        case 'C':
-            return `C ${formatNumber(command.c1.x)} ${formatNumber(command.c1.y)} ${formatNumber(command.c2.x)} ${formatNumber(
-                command.c2.y
-            )} ${formatNumber(command.to.x)} ${formatNumber(command.to.y)}`;
-        case 'Z':
-            return 'Z';
-    }
-};
-
+/** Narrow raw commands back to the small SVG subset used by the structured path model. */
 const isLineTo = (command: PathCommand): command is LineTo => command.cmd === 'L';
 const isCubicTo = (command: PathCommand): command is CubicTo => command.cmd === 'C';
 const isClosePath = (command: PathCommand): command is ClosePath => command.cmd === 'Z';
 const isLineOnlyOpenPath = (
     commands: readonly [MoveTo, OpenPathDrawCommand, ...OpenPathDrawCommand[]]
 ): commands is readonly [MoveTo, LineTo, ...LineTo[]] => commands.slice(1).every(isLineTo);
-
-const makeBasePath = <TKind extends Path['kind'], TCommands extends readonly PathCommand[]>(
-    kind: TKind,
-    commands: TCommands
-) =>
-    ({
-        kind,
-        commands,
-        d: commands.map(serializeCommand).join(' '),
-    }) as Extract<Path, { kind: TKind }> & { commands: TCommands };
-
-export const makeLinearPath = (start: PathPoint, end: PathPoint): LinearPath =>
-    makeBasePath('ml', [moveTo(start), lineTo(end)] as const);
-
-export const makeSharpTurnPath = (start: PathPoint, corner: PathPoint, end: PathPoint): SharpTurnPath =>
-    makeBasePath('mll', [moveTo(start), lineTo(corner), lineTo(end)] as const);
-
-export const makeRoundedTurnPath = (
-    start: PathPoint,
-    lineEnd: PathPoint,
-    c1: PathPoint,
-    c2: PathPoint,
-    curveEnd: PathPoint,
-    end: PathPoint
-): RoundedTurnPath =>
-    makeBasePath('mlcl', [moveTo(start), lineTo(lineEnd), cubicTo(c1, c2, curveEnd), lineTo(end)] as const);
-
-const makeComplexOpenPath = (
-    commands: readonly [MoveTo, OpenPathDrawCommand, OpenPathDrawCommand, ...OpenPathDrawCommand[]]
-): ComplexOpenPath => makeBasePath('complex-open', commands);
-
-export const makeClosedAreaPath = (
-    commands: readonly [MoveTo, OpenPathDrawCommand, OpenPathDrawCommand, ...OpenPathDrawCommand[], ClosePath]
-): ClosedAreaPath => makeBasePath('closed-area', commands);
-
-export const makeClosedAreaPathFromOpenCommands = (
-    commands: readonly [MoveTo, OpenPathDrawCommand, OpenPathDrawCommand, ...OpenPathDrawCommand[]]
-): ClosedAreaPath => makeClosedAreaPath([...commands, closePath()] as const);
 
 /**
  * Reconstruct the narrowest path kind from a command list.
@@ -203,6 +86,7 @@ const arePointsCollinear = (points: readonly PathPoint[]) => {
     return points.every(point => Math.abs((point.x - anchor.x) * dy - (point.y - anchor.y) * dx) <= 1e-9);
 };
 
+/** Extract numeric operands from a serialized SVG path and validate the expected arity eagerly. */
 const extractPathNumbers = (pathD: string, expectedCount: number): number[] => {
     const numbers = pathD.match(/-?(?:\d+\.?\d*|\.\d+)(?:e[-+]?\d+)?/gi)?.map(Number) ?? [];
     if (numbers.length !== expectedCount || numbers.some(value => !Number.isFinite(value))) {
@@ -224,8 +108,10 @@ export const parseRoundedTurnPath = (pathD: string): RoundedTurnPath => {
     );
 };
 
+/** Every supported path starts with `M`, so the first command endpoint is always the start point. */
 export const getStartPoint = (path: Path): PathPoint => path.commands[0].to;
 
+/** For closed areas, ignore the trailing `Z` and return the last drawable endpoint instead. */
 export const getEndPoint = (path: Path): PathPoint => {
     const lastCommand = path.kind === 'closed-area' ? path.commands.at(-2) : path.commands.at(-1);
     if (!lastCommand || isClosePath(lastCommand)) {
@@ -257,13 +143,16 @@ export const concatOpenPaths = (paths: readonly [OpenPath, ...OpenPath[]] | read
     return makeOpenPathFromCommands(commands as [MoveTo, OpenPathDrawCommand, ...OpenPathDrawCommand[]]);
 };
 
+/** Split a straight segment at its midpoint for styles that render each half independently. */
 export const splitLinearPath = (path: LinearPath): [LinearPath, LinearPath] => {
     const [start, end] = [path.commands[0].to, path.commands[1].to];
     const middle = makePoint((start.x + end.x) / 2, (start.y + end.y) / 2);
     return [makeLinearPath(start, middle), makeLinearPath(middle, end)];
 };
 
+/** Small helper for style code that only handles straight segments. */
 export const isLinearPath = (path: OpenPath): path is LinearPath => path.kind === 'ml';
 
+/** Short open paths have dedicated render/offset logic and are worth detecting explicitly. */
 export const isShortOpenPath = (path: OpenPath): path is ShortOpenPath =>
     path.kind === 'ml' || path.kind === 'mll' || path.kind === 'mlcl';
